[REFACTORING] Relocate SessionRoutes in toutes package
