+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/routes/SessionRoutesTest.scala
@@ -17,7 +17,7 @@
  * under the License.                                           *
  * ***************************************************************/
 
-package org.apache.james.jmap.http
+package org.apache.james.jmap.routes
 
 import java.nio.charset.StandardCharsets
 
@@ -31,9 +31,10 @@ import net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson
 import org.apache.http.HttpStatus
 import org.apache.james.core.Username
 import org.apache.james.jmap._
-import org.apache.james.jmap.http.SessionRoutesTest.{BOB, TEST_CONFIGURATION}
+import org.apache.james.jmap.http.Authenticator
 import org.apache.james.jmap.model.JmapRfc8621Configuration
 import org.apache.james.jmap.model.JmapRfc8621Configuration.LOCALHOST_URL_PREFIX
+import org.apache.james.jmap.routes.SessionRoutesTest.{BOB, TEST_CONFIGURATION}
 import org.apache.james.mailbox.MailboxSession
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito._
