+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/routes/SessionSupplierTest.scala
@@ -17,11 +17,11 @@
  * under the License.                                           *
  * ***************************************************************/
 
-package org.apache.james.jmap.http
+package org.apache.james.jmap.routes
 
 import org.apache.james.core.Username
-import org.apache.james.jmap.http.SessionSupplierTest.USERNAME
 import org.apache.james.jmap.model.JmapRfc8621Configuration
+import org.apache.james.jmap.routes.SessionSupplierTest.USERNAME
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
